Decompose MDL shader cache translation unit
Extract key construction, material-model generation, parameter binding,
and texture-graph generation into focused components. Leave
MdlShaderCache.cpp responsible only for cache orchestration.

Move MdlTextureGraphGenerator behind a private implementation and place
each component interface in its corresponding header.
